requirements.txt
asterisk-ami
pyst2
requests
python-dotenv
